feat(16096): Multi-arity ops
